Move proof verification to utility methods
